move exchange_helpers to data module
