+++ freqtrade/tests/data/test_convert.py
@@ -1,7 +1,7 @@
 # pragma pylint: disable=missing-docstring, C0103
 import logging
 
-from freqtrade.exchange.exchange_helpers import parse_ticker_dataframe
+from freqtrade.data.convert import parse_ticker_dataframe
 from freqtrade.tests.conftest import log_has
